refs CCDEV-16 Add default empty list.
Fixed issue when pci devices list is not defined
None type can't be extended. So it is now set to
empty list as a default value.

Removed service dependency from the configuration
now if you don't state that nvgrid is available
you can still run tests.

If you set nvgrid as available service some
defaults would be applied to the test config. It
is used to verify nvidia videocard presence.