Replace lsof with explicit modinfo
At the moment we verify existance of modules in gues system through
lsmod. Though, lsmod does not output buit-in drivers - only extra
kernel modules.

In order to unify approaches and account for built-in drivers, we
replace lsof usage with a generator, which will execute modinfo for
each module we need to check for.